# Improve performance of some Time methods